Update to new OBEX transport and SDP search
The fetchMasInstances() search API have been replaced with a more
generic sdpSearch(UUID) API, and therefore the constructor of
BluetoothMasClient should take in a SdpMasRecord in stead of
a BluetoothMasInstance.

Any app using this library must use the new sdpSearch API and
listen for the new ACTION_SDP_RECORD intent.

Change-Id: I4a9a097fc05e227f606b8eb127a2762ec595cbab
Signed-off-by: Casper Bonde <c.bonde@samsung.com>